+++ packages/core/src/test-script/WebpackCompiler.ts
@@ -2,10 +2,17 @@ import { Compiler, CompilerOutput } from '@flood/element-compiler'
 import parseComments from 'comment-parser'
 import { ITestScript } from '../ITestScript'
 import { VMScript } from 'vm2'
-import { TestScriptOptions, TestScriptDefaultOptions, TestScriptError } from '../TestScript'
+import { TestScriptError } from '../TestScriptError'
+import { TestScriptOptions } from '../TestScriptOptions'
 import { readFileSync } from 'fs-extra'
 import { SourceUnmapper } from './SourceUnmapper'
 import { dirname } from 'path'
+
+// FIXME: WebpackCompiler currently doesn't do anything with this, but it should
+export const TestScriptDefaultOptions: TestScriptOptions = {
+	stricterTypeChecking: false,
+	traceResolution: false,
+}
 
 export default class WebpackCompiler implements ITestScript {
 	private vmScriptCache: VMScript