+++ packages/element-api/CHANGELOG.md
@@ -3,6 +3,155 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2-beta.2](https://github.com/flood-io/element/compare/v1.2.2-beta.1...v1.2.2-beta.2) (2020-04-02)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.2.2-beta.0](https://github.com/flood-io/element/compare/v1.2.1...v1.2.2-beta.0) (2020-04-02)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.2.1](https://github.com/flood-io/element/compare/v1.2.1-feat.0...v1.2.1) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## 1.2.1-feat.0 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+# [1.2.0](https://github.com/flood-io/element/compare/v1.2.0-feat.2...v1.2.0) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+# 1.2.0-feat.2 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+# 1.2.0-feat.1 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+# 1.2.0-feat.0 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+# 1.1.0-feat.0 (2020-04-01)
+
+
+### Features
+
+* **example:** tesing release version bump to 1.1 ([5dbc19f](https://github.com/flood-io/element/commit/5dbc19fa8ac2dbaabe769786d69da95c8186718f))
+
+
+
+
+
+## 1.0.11-feat.4 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## 1.0.11-feat.3 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## 1.0.11-feat.2 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## 1.0.11-feat.1 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## 1.0.11-feat.0 (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.0.11-alpha.2](https://github.com/flood-io/element/compare/v1.0.10-beta.63...v1.0.11-alpha.2) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.0.11-alpha.1](https://github.com/flood-io/element/compare/v1.0.10-beta.63...v1.0.11-alpha.1) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.0.11-alpha.0](https://github.com/flood-io/element/compare/v1.0.10-beta.63...v1.0.11-alpha.0) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.0.10](https://github.com/flood-io/element/compare/v1.0.10-beta.63...v1.0.10) (2020-04-01)
+
+**Note:** Version bump only for package @flood/element-api
+
+
+
+
+
+## [1.0.10-beta.63](https://github.com/flood-io/element/compare/v1.0.10-beta.62...v1.0.10-beta.63) (2020-04-01)
+
 ## [1.0.10](https://github.com/flood-io/element/compare/v1.0.10-beta.62...v1.0.10) (2020-03-27)
 
 **Note:** Version bump only for package @flood/element-api